Consolidate FragmentInfo for ViewPager
