+++ MvvmCross-AndroidSupport/MvvmCross.Droid.Support.Fragment/MvvmCross.Droid.Support.Fragment.csproj
@@ -83,6 +83,7 @@
     <Compile Include="MvxFragmentStatePagerAdapter.cs" />
     <Compile Include="MvxFragmentViewExtensionMethods.cs" />
     <Compile Include="MvxTabsFragmentActivity.cs" />
+    <Compile Include="MvxViewPagerFragment.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config" />
